+++ dicom_validator/validator/iod_validator.py
@@ -18,7 +18,7 @@ class IODValidator(object):
         self._module_info = module_info
         self._dict_info = dict_info
         self.errors = {}
-        self.logger = logging.getLogger()
+        self.logger = logging.getLogger('validator')
         self.logger.level = log_level
         if not self.logger.handlers:
             self.logger.addHandler(logging.StreamHandler(sys.stdout))